Added the offline installer code under install helpers
THe offline installer is a small script ran on a computer with connectivity to the internet that dowloads all zoomdata and dependencies.  It zips these up in a directory structure that is decompressed on target install host were a local repo is built.  Zoom can then be installed with just a few changes to the primary install script.